onedpl: renamed from parallelstl
